pyproject.toml
[project]
name = "trading212-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hishel>=0.1.2",
    "httpx>=0.28.1",
    "mcp[cli]>=1.8.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
]
